+++ src/pages/PostsManagerPage.tsx
@@ -2,29 +2,37 @@ import { useEffect, useState } from "react"
 import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
 import { useLocation, useNavigate } from "react-router-dom"
 import {
+  Dialog,
+  DialogContents,
+  DialogHeader,
+  DialogTitle,
+} from "../shared/ui/Dialog"
+import {
   Button,
+} from "../shared/ui/Button/Button"
+import {
+  Input
+} from "../shared/ui/InputBox/InputBox"
+
+import {  Table,
+TableBody,
+TableCell,
+TableHead,
+TableHeader,
+TableRow} from "../shared/ui/Table"
+import {
   Card,
   CardContent,
   CardHeader,
-  CardTitle,
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  Input,
+  CardTitle
+} from "../shared/ui/Card"
+import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
-  SelectValue,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-  Textarea,
-} from "../shared/ui"
+  SelectValue} from "../shared/ui/Select"
+import {Textarea } from "../shared/ui/Textarea/Textarea"
 
 const PostsManager = () => {
   const navigate = useNavigate()
@@ -32,8 +40,8 @@ const PostsManager = () => {
   const queryParams = new URLSearchParams(location.search)
 
   // 상태 관리
-  const [posts, setPosts] = useState([])
-  const [total, setTotal] = useState(0)
+  // const [posts, setPosts] = useState([])
+  // const [total, setTotal] = useState(0)
   const [skip, setSkip] = useState(parseInt(queryParams.get("skip") || "0"))
   const [limit, setLimit] = useState(parseInt(queryParams.get("limit") || "10"))
   const [searchQuery, setSearchQuery] = useState(queryParams.get("search") || "")
@@ -43,10 +51,10 @@ const PostsManager = () => {
   const [showAddDialog, setShowAddDialog] = useState(false)
   const [showEditDialog, setShowEditDialog] = useState(false)
   const [newPost, setNewPost] = useState({ title: "", body: "", userId: 1 })
-  const [loading, setLoading] = useState(false)
-  const [tags, setTags] = useState([])
+  // const [loading, setLoading] = useState(false)
+  // const [tags, setTags] = useState([])
   const [selectedTag, setSelectedTag] = useState(queryParams.get("tag") || "")
-  const [comments, setComments] = useState({})
+  // const [comments, setComments] = useState({})
   const [selectedComment, setSelectedComment] = useState(null)
   const [newComment, setNewComment] = useState({ body: "", postId: null, userId: 1 })
   const [showAddCommentDialog, setShowAddCommentDialog] = useState(false)
@@ -67,221 +75,221 @@ const PostsManager = () => {
     navigate(`?${params.toString()}`)
   }
 
-  // 게시물 가져오기
-  const fetchPosts = () => {
-    setLoading(true)
-    let postsData
-    let usersData
+  // // 게시물 가져오기
+  // const fetchPosts = () => {
+  //   setLoading(true)
+  //   let postsData
+  //   let usersData
 
-    fetch(`/api/posts?limit=${limit}&skip=${skip}`)
-      .then((response) => response.json())
-      .then((data) => {
-        postsData = data
-        return fetch("/api/users?limit=0&select=username,image")
-      })
-      .then((response) => response.json())
-      .then((users) => {
-        usersData = users.users
-        const postsWithUsers = postsData.posts.map((post) => ({
-          ...post,
-          author: usersData.find((user) => user.id === post.userId),
-        }))
-        setPosts(postsWithUsers)
-        setTotal(postsData.total)
-      })
-      .catch((error) => {
-        console.error("게시물 가져오기 오류:", error)
-      })
-      .finally(() => {
-        setLoading(false)
-      })
-  }
+  //   fetch(`/api/posts?limit=${limit}&skip=${skip}`)
+  //     .then((response) => response.json())
+  //     .then((data) => {
+  //       postsData = data
+  //       return fetch("/api/users?limit=0&select=username,image")
+  //     })
+  //     .then((response) => response.json())
+  //     .then((users) => {
+  //       usersData = users.users
+  //       const postsWithUsers = postsData.posts.map((post) => ({
+  //         ...post,
+  //         author: usersData.find((user) => user.id === post.userId),
+  //       }))
+  //       setPosts(postsWithUsers)
+  //       setTotal(postsData.total)
+  //     })
+  //     .catch((error) => {
+  //       console.error("게시물 가져오기 오류:", error)
+  //     })
+  //     .finally(() => {
+  //       setLoading(false)
+  //     })
+  // }
 
-  // 태그 가져오기
-  const fetchTags = async () => {
-    try {
-      const response = await fetch("/api/posts/tags")
-      const data = await response.json()
-      setTags(data)
-    } catch (error) {
-      console.error("태그 가져오기 오류:", error)
-    }
-  }
+  // // 태그 가져오기
+  // const fetchTags = async () => {
+  //   try {
+  //     const response = await fetch("/api/posts/tags")
+  //     const data = await response.json()
+  //     setTags(data)
+  //   } catch (error) {
+  //     console.error("태그 가져오기 오류:", error)
+  //   }
+  // }
 
-  // 게시물 검색
-  const searchPosts = async () => {
-    if (!searchQuery) {
-      fetchPosts()
-      return
-    }
-    setLoading(true)
-    try {
-      const response = await fetch(`/api/posts/search?q=${searchQuery}`)
-      const data = await response.json()
-      setPosts(data.posts)
-      setTotal(data.total)
-    } catch (error) {
-      console.error("게시물 검색 오류:", error)
-    }
-    setLoading(false)
-  }
+  // // 게시물 검색
+  // const searchPosts = async () => {
+  //   if (!searchQuery) {
+  //     fetchPosts()
+  //     return
+  //   }
+  //   setLoading(true)
+  //   try {
+  //     const response = await fetch(`/api/posts/search?q=${searchQuery}`)
+  //     const data = await response.json()
+  //     setPosts(data.posts)
+  //     setTotal(data.total)
+  //   } catch (error) {
+  //     console.error("게시물 검색 오류:", error)
+  //   }
+  //   setLoading(false)
+  // }
 
-  // 태그별 게시물 가져오기
-  const fetchPostsByTag = async (tag) => {
-    if (!tag || tag === "all") {
-      fetchPosts()
-      return
-    }
-    setLoading(true)
-    try {
-      const [postsResponse, usersResponse] = await Promise.all([
-        fetch(`/api/posts/tag/${tag}`),
-        fetch("/api/users?limit=0&select=username,image"),
-      ])
-      const postsData = await postsResponse.json()
-      const usersData = await usersResponse.json()
+  // // 태그별 게시물 가져오기
+  // const fetchPostsByTag = async (tag) => {
+  //   if (!tag || tag === "all") {
+  //     fetchPosts()
+  //     return
+  //   }
+  //   setLoading(true)
+  //   try {
+  //     const [postsResponse, usersResponse] = await Promise.all([
+  //       fetch(`/api/posts/tag/${tag}`),
+  //       fetch("/api/users?limit=0&select=username,image"),
+  //     ])
+  //     const postsData = await postsResponse.json()
+  //     const usersData = await usersResponse.json()
 
-      const postsWithUsers = postsData.posts.map((post) => ({
-        ...post,
-        author: usersData.users.find((user) => user.id === post.userId),
-      }))
+  //     const postsWithUsers = postsData.posts.map((post) => ({
+  //       ...post,
+  //       author: usersData.users.find((user) => user.id === post.userId),
+  //     }))
 
-      setPosts(postsWithUsers)
-      setTotal(postsData.total)
-    } catch (error) {
-      console.error("태그별 게시물 가져오기 오류:", error)
-    }
-    setLoading(false)
-  }
+  //     setPosts(postsWithUsers)
+  //     setTotal(postsData.total)
+  //   } catch (error) {
+  //     console.error("태그별 게시물 가져오기 오류:", error)
+  //   }
+  //   setLoading(false)
+  // }
 
-  // 게시물 추가
-  const addPost = async () => {
-    try {
-      const response = await fetch("/api/posts/add", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(newPost),
-      })
-      const data = await response.json()
-      setPosts([data, ...posts])
-      setShowAddDialog(false)
-      setNewPost({ title: "", body: "", userId: 1 })
-    } catch (error) {
-      console.error("게시물 추가 오류:", error)
-    }
-  }
+  // // 게시물 추가
+  // const addPost = async () => {
+  //   try {
+  //     const response = await fetch("/api/posts/add", {
+  //       method: "POST",
+  //       headers: { "Content-Type": "application/json" },
+  //       body: JSON.stringify(newPost),
+  //     })
+  //     const data = await response.json()
+  //     setPosts([data, ...posts])
+  //     setShowAddDialog(false)
+  //     setNewPost({ title: "", body: "", userId: 1 })
+  //   } catch (error) {
+  //     console.error("게시물 추가 오류:", error)
+  //   }
+  // }
 
-  // 게시물 업데이트
-  const updatePost = async () => {
-    try {
-      const response = await fetch(`/api/posts/${selectedPost.id}`, {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(selectedPost),
-      })
-      const data = await response.json()
-      setPosts(posts.map((post) => (post.id === data.id ? data : post)))
-      setShowEditDialog(false)
-    } catch (error) {
-      console.error("게시물 업데이트 오류:", error)
-    }
-  }
+  // // 게시물 업데이트
+  // const updatePost = async () => {
+  //   try {
+  //     const response = await fetch(`/api/posts/${selectedPost.id}`, {
+  //       method: "PUT",
+  //       headers: { "Content-Type": "application/json" },
+  //       body: JSON.stringify(selectedPost),
+  //     })
+  //     const data = await response.json()
+  //     setPosts(posts.map((post) => (post.id === data.id ? data : post)))
+  //     setShowEditDialog(false)
+  //   } catch (error) {
+  //     console.error("게시물 업데이트 오류:", error)
+  //   }
+  // }
 
-  // 게시물 삭제
-  const deletePost = async (id) => {
-    try {
-      await fetch(`/api/posts/${id}`, {
-        method: "DELETE",
-      })
-      setPosts(posts.filter((post) => post.id !== id))
-    } catch (error) {
-      console.error("게시물 삭제 오류:", error)
-    }
-  }
+  // // 게시물 삭제
+  // const deletePost = async (id) => {
+  //   try {
+  //     await fetch(`/api/posts/${id}`, {
+  //       method: "DELETE",
+  //     })
+  //     setPosts(posts.filter((post) => post.id !== id))
+  //   } catch (error) {
+  //     console.error("게시물 삭제 오류:", error)
+  //   }
+  // }
 
-  // 댓글 가져오기
-  const fetchComments = async (postId) => {
-    if (comments[postId]) return // 이미 불러온 댓글이 있으면 다시 불러오지 않음
-    try {
-      const response = await fetch(`/api/comments/post/${postId}`)
-      const data = await response.json()
-      setComments((prev) => ({ ...prev, [postId]: data.comments }))
-    } catch (error) {
-      console.error("댓글 가져오기 오류:", error)
-    }
-  }
+  // // 댓글 가져오기
+  // const fetchComments = async (postId) => {
+  //   if (comments[postId]) return // 이미 불러온 댓글이 있으면 다시 불러오지 않음
+  //   try {
+  //     const response = await fetch(`/api/comments/post/${postId}`)
+  //     const data = await response.json()
+  //     setComments((prev) => ({ ...prev, [postId]: data.comments }))
+  //   } catch (error) {
+  //     console.error("댓글 가져오기 오류:", error)
+  //   }
+  // }
 
-  // 댓글 추가
-  const addComment = async () => {
-    try {
-      const response = await fetch("/api/comments/add", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(newComment),
-      })
-      const data = await response.json()
-      setComments((prev) => ({
-        ...prev,
-        [data.postId]: [...(prev[data.postId] || []), data],
-      }))
-      setShowAddCommentDialog(false)
-      setNewComment({ body: "", postId: null, userId: 1 })
-    } catch (error) {
-      console.error("댓글 추가 오류:", error)
-    }
-  }
+  // // 댓글 추가
+  // const addComment = async () => {
+  //   try {
+  //     const response = await fetch("/api/comments/add", {
+  //       method: "POST",
+  //       headers: { "Content-Type": "application/json" },
+  //       body: JSON.stringify(newComment),
+  //     })
+  //     const data = await response.json()
+  //     setComments((prev) => ({
+  //       ...prev,
+  //       [data.postId]: [...(prev[data.postId] || []), data],
+  //     }))
+  //     setShowAddCommentDialog(false)
+  //     setNewComment({ body: "", postId: null, userId: 1 })
+  //   } catch (error) {
+  //     console.error("댓글 추가 오류:", error)
+  //   }
+  // }
 
-  // 댓글 업데이트
-  const updateComment = async () => {
-    try {
-      const response = await fetch(`/api/comments/${selectedComment.id}`, {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ body: selectedComment.body }),
-      })
-      const data = await response.json()
-      setComments((prev) => ({
-        ...prev,
-        [data.postId]: prev[data.postId].map((comment) => (comment.id === data.id ? data : comment)),
-      }))
-      setShowEditCommentDialog(false)
-    } catch (error) {
-      console.error("댓글 업데이트 오류:", error)
-    }
-  }
+  // // 댓글 업데이트
+  // const updateComment = async () => {
+  //   try {
+  //     const response = await fetch(`/api/comments/${selectedComment.id}`, {
+  //       method: "PUT",
+  //       headers: { "Content-Type": "application/json" },
+  //       body: JSON.stringify({ body: selectedComment.body }),
+  //     })
+  //     const data = await response.json()
+  //     setComments((prev) => ({
+  //       ...prev,
+  //       [data.postId]: prev[data.postId].map((comment) => (comment.id === data.id ? data : comment)),
+  //     }))
+  //     setShowEditCommentDialog(false)
+  //   } catch (error) {
+  //     console.error("댓글 업데이트 오류:", error)
+  //   }
+  // }
 
-  // 댓글 삭제
-  const deleteComment = async (id, postId) => {
-    try {
-      await fetch(`/api/comments/${id}`, {
-        method: "DELETE",
-      })
-      setComments((prev) => ({
-        ...prev,
-        [postId]: prev[postId].filter((comment) => comment.id !== id),
-      }))
-    } catch (error) {
-      console.error("댓글 삭제 오류:", error)
-    }
-  }
+  // // 댓글 삭제
+  // const deleteComment = async (id, postId) => {
+  //   try {
+  //     await fetch(`/api/comments/${id}`, {
+  //       method: "DELETE",
+  //     })
+  //     setComments((prev) => ({
+  //       ...prev,
+  //       [postId]: prev[postId].filter((comment) => comment.id !== id),
+  //     }))
+  //   } catch (error) {
+  //     console.error("댓글 삭제 오류:", error)
+  //   }
+  // }
 
-  // 댓글 좋아요
-  const likeComment = async (id, postId) => {
-    try {
-      const response = await fetch(`/api/comments/${id}`, {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ likes: comments[postId].find((c) => c.id === id).likes + 1 }),
-      })
-      const data = await response.json()
-      setComments((prev) => ({
-        ...prev,
-        [postId]: prev[postId].map((comment) => (comment.id === data.id ? data : comment)),
-      }))
-    } catch (error) {
-      console.error("댓글 좋아요 오류:", error)
-    }
-  }
+  // // 댓글 좋아요
+  // const likeComment = async (id, postId) => {
+  //   try {
+  //     const response = await fetch(`/api/comments/${id}`, {
+  //       method: "PATCH",
+  //       headers: { "Content-Type": "application/json" },
+  //       body: JSON.stringify({ likes: comments[postId].find((c) => c.id === id).likes + 1 }),
+  //     })
+  //     const data = await response.json()
+  //     setComments((prev) => ({
+  //       ...prev,
+  //       [postId]: prev[postId].map((comment) => (comment.id === data.id ? data : comment)),
+  //     }))
+  //   } catch (error) {
+  //     console.error("댓글 좋아요 오류:", error)
+  //   }
+  // }
 
   // 게시물 상세 보기
   const openPostDetail = (post) => {
@@ -570,7 +578,7 @@ const PostsManager = () => {
 
       {/* 게시물 추가 대화상자 */}
       <Dialog open={showAddDialog} onOpenChange={setShowAddDialog}>
-        <DialogContent>
+        <DialogContents>
           <DialogHeader>
             <DialogTitle>새 게시물 추가</DialogTitle>
           </DialogHeader>
@@ -594,12 +602,12 @@ const PostsManager = () => {
             />
             <Button onClick={addPost}>게시물 추가</Button>
           </div>
-        </DialogContent>
+        </DialogContents>
       </Dialog>
 
       {/* 게시물 수정 대화상자 */}
       <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
-        <DialogContent>
+        <DialogContents>
           <DialogHeader>
             <DialogTitle>게시물 수정</DialogTitle>
           </DialogHeader>
@@ -617,12 +625,12 @@ const PostsManager = () => {
             />
             <Button onClick={updatePost}>게시물 업데이트</Button>
           </div>
-        </DialogContent>
+        </DialogContents>
       </Dialog>
 
       {/* 댓글 추가 대화상자 */}
       <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
-        <DialogContent>
+        <DialogContents>
           <DialogHeader>
             <DialogTitle>새 댓글 추가</DialogTitle>
           </DialogHeader>
@@ -634,12 +642,12 @@ const PostsManager = () => {
             />
             <Button onClick={addComment}>댓글 추가</Button>
           </div>
-        </DialogContent>
+        </DialogContents>
       </Dialog>
 
       {/* 댓글 수정 대화상자 */}
       <Dialog open={showEditCommentDialog} onOpenChange={setShowEditCommentDialog}>
-        <DialogContent>
+        <DialogContents>
           <DialogHeader>
             <DialogTitle>댓글 수정</DialogTitle>
           </DialogHeader>
@@ -651,12 +659,12 @@ const PostsManager = () => {
             />
             <Button onClick={updateComment}>댓글 업데이트</Button>
           </div>
-        </DialogContent>
+        </DialogContents>
       </Dialog>
 
       {/* 게시물 상세 보기 대화상자 */}
       <Dialog open={showPostDetailDialog} onOpenChange={setShowPostDetailDialog}>
-        <DialogContent className="max-w-3xl">
+        <DialogContents className="max-w-3xl">
           <DialogHeader>
             <DialogTitle>{highlightText(selectedPost?.title, searchQuery)}</DialogTitle>
           </DialogHeader>
@@ -664,12 +672,12 @@ const PostsManager = () => {
             <p>{highlightText(selectedPost?.body, searchQuery)}</p>
             {renderComments(selectedPost?.id)}
           </div>
-        </DialogContent>
+        </DialogContents>
       </Dialog>
 
       {/* 사용자 모달 */}
       <Dialog open={showUserModal} onOpenChange={setShowUserModal}>
-        <DialogContent>
+        <DialogContents>
           <DialogHeader>
             <DialogTitle>사용자 정보</DialogTitle>
           </DialogHeader>
@@ -698,7 +706,7 @@ const PostsManager = () => {
               </p>
             </div>
           </div>
-        </DialogContent>
+        </DialogContents>
       </Dialog>
     </Card>
   )
