+++ src/shared/ui/Card/Card.tsx
@@ -1,0 +1,6 @@
+import { forwardRef } from "react"
+
+export const Card = forwardRef(({ className, ...props }, ref) => (
+  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+))
+Card.displayName = "Card"
