+++ src/shared/ui/Card/CardContent.tsx
@@ -1,0 +1,6 @@
+import { forwardRef } from "react"
+
+export const CardContent = forwardRef(({ className, ...props }, ref) => (
+  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
+))
+CardContent.displayName = "CardContent"
