+++ src/shared/ui/Card/CardHeader.tsx
@@ -1,0 +1,6 @@
+import { forwardRef } from "react"
+
+export const CardHeader = forwardRef(({ className, ...props }, ref) => (
+  <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
+))
+CardHeader.displayName = "CardHeader"
