+++ src/shared/ui/Dialog/Dialog.tsx
@@ -1,0 +1,9 @@
+// Dialog.tsx
+import * as DialogPrimitive from '@radix-ui/react-dialog';
+
+export const Dialog = DialogPrimitive.Root;
+export const DialogTrigger = DialogPrimitive.Trigger;
+export const DialogPortal = DialogPrimitive.Portal;
+export const DialogOverlay = DialogPrimitive.Overlay;
+export const DialogContent = DialogPrimitive.Content;
+export const DialogClose = DialogPrimitive.Close;
