+++ src/shared/ui/Dialog/DialogTitle.tsx
@@ -1,0 +1,11 @@
+import { forwardRef } from "react"
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+
+export const DialogTitle = forwardRef(({ className, ...props }, ref) => (
+  <DialogPrimitive.Title
+    ref={ref}
+    className={`text-lg font-semibold leading-none tracking-tight ${className}`}
+    {...props}
+  />
+))
+DialogTitle.displayName = DialogPrimitive.Title.displayName
