+++ src/shared/ui/Dialog/index.ts
@@ -1,0 +1,4 @@
+export * from "./DialogContents"
+export * from "./DialogHeader"
+export * from "./DialogTitle"
+export * from "./Dialog"
