+++ src/shared/ui/Select/index.ts
@@ -1,0 +1,4 @@
+export * from "./SelectContent"
+export * from "./SelectItem"
+export * from "./SelectTrigger"
+export {Select,SelectValue,SelectGroup} from "./Select"
