+++ vite.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
         target: "https://dummyjson.com",
         changeOrigin: true,
         rewrite: (path) => path.replace(/^\/api/, ""),
+        secure: false 
       },
     },
   },
